add support for incomplete chunks
When receiving large events from the server the reader will sometimes
return incomplete chunks. The previous implementation did not fully
account for this and emitted incomplete messages / failed depending on
what `data-reader` option was provided.

The addition of `missionary` could probably be avoided but given the
stream nature of this I felt like it might be a good idea to use it.