Add a failing test for chunks not including the event boundary
When (.read reader) returns a chunk that doesn't include an
event delimiter, we'd expect it to be buffered and parsed
in subsequent calls to .read when we get the rest of the event.
However, this doesn't currently work due to a bug.

This commit adds a failing test for that case. Note that the
test would have worked if the first chunk ended with \r\n\r\n.